Move draw selection window to selection manager
